@@ -61,8 +61,9 @@
     "@types/react-dom": "^17.0.11",
     "ahooks": "^3.1.9",
     "antd": "^4.17.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "normalize.css": "^8.0.1",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-router-dom": "^6.2.1"
   },
   "browserslist": [
