@@ -8,10 +8,11 @@ import './App.scss';
 
 // import { hot } from 'react-hot-loader/root';
 // import Layout from '@layout/index';
+import Layout from '@examples/someHooks';
 // import Layout from '@examples/mingming';
 // import Layout from './remainingTime/index';
 // import Layout from '@examples/canvas';
-import Layout from '@examples/videoCarousel';
+// import Layout from '@examples/videoCarousel';
 // import Layout from '@examples/designPatterns';
 // import Layout from '@examples/draggableElement';
 
